Change rbusDateTime_t to use fixed size structures
rbusDateTime_t relies on the platform int size in the underlying
structures. This causes faults when mixing lib32 and lib64 rbus
clients.

This change modifies the underlying definition to use fixed size
data fields, and adds conversion routines to move that to/from
the underlying platform 'struct tm' representation.

This also replaces the printf() and scanf() format specifiers with
PRId64 where necessary to avoid compiler warnings.