RDKB-45280:Update the rtLog to support RDKLogger appropriately and modify rbus and rbuscore to use the rtLog Macro (#35)
* *RDKB-45280:Update the rtLog to support RDKLogger appropriately and modify rbus and rbuscore to use the rtLog Macro

Reason for change:Update the rtLog to support RDKLogger appropriately and modify rbus and rbuscore to use the rtLog Macro
Test Procedure: Tested on VM: Run-rtrouted for RTMESSAGE log (Terminal1), RbusProvider(Term2) and rbuscli or rbusConsumer For RBUS logs(Term3), rbus_session_mgr For RBUSCORE logs(Term4)
Risks: Low
Priority: P1

Signed-off-by: Deepak <deepak_m@comcast.com>
Co-authored-by: Karunakaran A <48997923+karuna2git@users.noreply.github.com>

RDKC-12563,RDKB-45841 - Continuous flooding seen in various logs from RBUS module (#52)

* RDKC-12563: Continuous flooding seen in various logs from RTMESSAGE and RBUS module
* RDKB-45841 : Log files are empty after 5 min uptime after upgrade/reboot/FR

Reason for change: Change for RDK logger for we used the rtLogLevel_t as is to pass to RDKLogger for rdk_LogLevel which is wrong, RT_LOG_DEBUG is 0 where as RDK_LOG_FATAL is also 0. So this lead all the debug to be printed as FATAL and similarly, all the INFO to be printed as ERROR. Test Procedure: Trigger RDKC and RDKB Build and test by flashing respective images and verify Logs of on Device. Risks: Med
Priority: P0