RDKB-50186 : [On rbusDirect] Implementation of RBUS Publish API to send Raw Data
Reason for change: Changes made with rawdata API to support both direct mode and
normal connection. If normal connection exist, the data will be send through it
and if direct connection exist, the data will be send through it.
Test Procedure: Test and verified with rbuscli
Risks: Medium
Priority: P1
Signed-off-by: Gururaaja ESR <gururaja_erodesriranganramlingham@comcast.com>