TCXB8-2598 : Onewifi crash observed with function strlen, rtLogPrintf
Reason for change: Subscriptions list maintained in the provider side is not
mutex protected. Because of which when some component subscribe/unsubscribe
the event, which in turn affects the subscription list and at the same time
if provider tries to access the subscriptions list in rbusEvent_Publish
it leads to a crash. Mutex protection added for subscription list to avoid
the SIGSEGV.
Test Procedure: Test and verified
Risks: Medium
Priority: P1
Signed-off-by: Gururaaja ESR <gururaja_erodesriranganramlingham@comcast.com>