RDKB-51760 - Provider crash with multiple rbus_open() call
Reason for change: With multiple rbus_open() call with same component name,
we are closing and opening the handle again. After closing the handle,
we are not performing rbusConfig_CreateOnce(), because of which in rbus_regDataElements()
where we are calling rbusConfig_Get()->tmpDir leads to a NULL pointer deferencing.
Made changes to call _rbus_open_pre_initialize() in the issue scenario.
Test Procedure: Test and verified
Risks: Medium
Priority: P1
Signed-off-by: Gururaaja ESR <gururaja_erodesriranganramlingham@comcast.com>