Reason for change: When a dmcli query is made with a large number of dots, rtrouted crashes. For example: dmcli eRT getv Device.NAT.X_CISCO_COM_DMZ.InternalIP
The reason is that the tokenizeExpression function splits the datamodel path based on dots and writes into the workTokens array without checking the number of elements being written.
This causes a buffer overflow if the number of dots is more than the size of the workTokens array (32).
[Solution]
Adding expression validation at the time of routeAdd to safe return with Invalid Expression Error.
Test Procedure: Sanity.
Risks: None.

Signed-off-by: Vivek Narwaria Vivek_Narwaria@comcast.com