RDKC-15751: [Xcam2] Update WiFi connection not working via XFi app
Reason for change:
The button press event was not sent to xcam2 because the `rtsend` binary was missing in xw4.

Test Procedure:
Press the push button on Xw4 for 10 seconds and release. It should blink white and
the event receive log should be seen in `prvnmgr` logs.

Risks: High
Priority: P1

Signed-off-by: msarav043 <mariammal_saravanan@comcast.com>