RDKB-59999 : Drop Incoming Request when Provider is NOT Responding/Blocked (#269)
Reason for change: Drop Incoming Request when Provider is NOT Responding/Blocked. When the request posted to the connection/callback_thread but if Provider is NOT consuming, the messages are queued up and eventually it leads to out-of-memory for the whole system. So It is decided to DROP all the incoming messages when the Queue size is 25.
Test Procedure: Added Test Programs to simulate
Priority: P0
Risks: High
Signed-off-by: dshett549 <DEEPTHICHANDRASHEKAR_SHETTY@comcast.com>
Co-authored-by: Karunakaran A <48997923+karuna2git@users.noreply.github.com>