core: fixup null termination in rbuscore
The code changes in this PR were automatically generated by the
Permanence AI Coder and reviewed by @jweese and @eelenberg. This moves
a call from `strcpy` to `strncpy` and also takes care to null terminate
in the case where `strncpy` could hit its limit. Note that the use of
`sizeof` to calculate the max size of each buffer is valid here, since
both `m_privateDML` and `m_providerName` are arrays of known size.

on-behalf-of: @permanence-ai <github-ai@permanence.ai>