rbus: fixup null pointer dereference in _subscribe_async_callback_handler
The changes in this PR were automatically generated by the Permanence AI
Coder and reviewed by @dnovick and @jweese. The code was checking to make sure
`subscription` is non-NULL after it had already been dereferenced to get
its `asyncHandler` and `handle`. This change moves the use of `subscription`
inside the NULL check. An error message is logged in case `subscription`
is NULL.

on-behalf-of: @permanence-ai <github-ai@permanence.ai>